split model logic into 3 models
